pyproject.toml
[tool.poetry]
name = "wmcompanion"
version = "0.1.0"
description = "wmcompanion is a utility for connecting system events to user actions"
authors = ["Daniel Pereira <daniel@garajau.com.br>"]

[tool.poetry.dependencies]
python = "^3.10"
dbus-next = "^0.2"
systemd-python = { git = "https://github.com/systemd/python-systemd", rev = "b109b75" }

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
wmcompanion = "wmcompanion.cli:main"

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"
